fix loi predict

+++ model.py
@@ -1,6 +1,5 @@
 import torch.nn as nn
 
-
 class LSTMClassifier(nn.Module):
     """
     This is the simple RNN model we will be using to perform Sentiment Analysis.
@@ -16,7 +15,7 @@ class LSTMClassifier(nn.Module):
         self.lstm = nn.LSTM(embedding_dim, hidden_dim)
         self.dense = nn.Linear(in_features=hidden_dim, out_features=1)
         self.sig = nn.Sigmoid()
-
+        
         self.word_dict = None
 
     def forward(self, x):
@@ -24,10 +23,10 @@ class LSTMClassifier(nn.Module):
         Perform a forward pass of our model on some input.
         """
         x = x.t()
-        lengths = x[0, :]
-        reviews = x[1:, :]
+        lengths = x[0,:]
+        reviews = x[1:,:]
         embeds = self.embedding(reviews)
         lstm_out, _ = self.lstm(embeds)
         out = self.dense(lstm_out)
-        out = out[lengths - 1, range(len(lengths))]
+        out = out[(lengths - 1).long(), range(len(lengths))]
         return self.sig(out.squeeze())
